+++ tests/test_wrapper.py
@@ -23,7 +23,8 @@ import tempfile
 import unittest
 from unittest import mock
 
-import git_command
+import utils_for_test
+
 import main
 import wrapper
 
@@ -408,18 +409,7 @@ class GitCheckoutTestCase(RepoWrapperTestCase):
         remote = os.path.join(cls.GIT_DIR, "remote")
         os.mkdir(remote)
 
-        # Tests need to assume, that main is default branch at init,
-        # which is not supported in config until 2.28.
-        if git_command.git_require((2, 28, 0)):
-            initstr = "--initial-branch=main"
-        else:
-            # Use template dir for init.
-            templatedir = tempfile.mkdtemp(prefix=".test-template")
-            with open(os.path.join(templatedir, "HEAD"), "w") as fp:
-                fp.write("ref: refs/heads/main\n")
-            initstr = "--template=" + templatedir
-
-        run_git("init", initstr, cwd=remote)
+        utils_for_test.init_git_tree(remote)
         run_git("commit", "--allow-empty", "-minit", cwd=remote)
         run_git("branch", "stable", cwd=remote)
         run_git("tag", "v1.0", cwd=remote)
